bug fix in /myaccount and add some simple GetMappings in Product and Address Controller

+++ backend/spring-boot-ecommerce-website/src/main/java/com/ecommerce/ecommercewebsite/service/AddressService.java
@@ -5,6 +5,6 @@ import com.ecommerce.ecommercewebsite.entity.Address;
 import java.util.List;
 
 public interface AddressService {
-    List<Address> getAllAddressesByUserId(int user_id);
+    List<Address> getAllAddressesByUserId(Long user_id);
     void insertAddress(Long user_id, String city, String postal_code, String street, String country, int street_number);
 }
