+++ backend/spring-boot-ecommerce-website/src/main/java/com/ecommerce/ecommercewebsite/service/AddressServiceImpl.java
@@ -16,7 +16,7 @@ public class AddressServiceImpl implements AddressService{
     }
 
     @Override
-    public List<Address> getAllAddressesByUserId(int user_id) {
+    public List<Address> getAllAddressesByUserId(Long user_id) {
         return addressRepository.getAllAddressesByUserId(user_id);
     }
